feat: GameTime 객체 도입으로 원시타입 포장
